feat(validation): introduce form and widget level validation
We have a Validator object from Laravel (made with the Facade) which we
then use with the form values we pack() up.
That produces an object we can query to make sure we can display errors
in the form. So we now have leveraged the Laravel Validation rules into
our custom widgets.