feat(initalvals): Initial vals, Widget::changed()
We want widgets to be given a value, then tell us if they have been
changed and if so then tell us what columns to change, so we need to
track initial values, and entered values. Usually this is easy, but with
PasswordCreate we have the "unchanged" password appear in a Form post as
an empty string, rather than, say a "name" field, where the name comes
back again.

Having the widget in control of which columns it needs, and can set
could lead to namespace issues, but it also allows for a single widget
to do some clever stuff and store it's data over several columns, or
even related tables, which we're going to need here