feat(checkbox): Added checkbox widget
This change required a change in the design of how forms work. We want
each widget to know what it's value is, inspect the request to see what
data has come back and figure out what has changed. With most widgets
this is just looking at the value, but for a checkbox, an unticked
widget doesn't appear in the HTTP request. So we now need to understand
where the data is being loaded from and potentially change the widget
logic based on that. So we now no longer have populate() but a variety
of populateFrom* functions so context is preserved through to the
widget.